Rename VeRDD to VeRDDOps
