+++ src/main/scala/com/nec/spark/planning/plans/SparkToVectorEnginePlan.scala
@@ -2,22 +2,18 @@ package com.nec.spark.planning.plans
 
 import com.nec.arrow.ArrowEncodingSettings
 import com.nec.cache.{ColumnarBatchToVeColBatch, CycloneCacheBase, DualMode}
-import com.nec.spark.{SparkCycloneExecutorPlugin, planning}
+import com.nec.spark.SparkCycloneExecutorPlugin
 import com.nec.spark.planning.plans.SparkToVectorEnginePlan.ConvertColumnarToColumnar
 import com.nec.spark.planning.{DataCleanup, PlanMetrics, SupportsVeColBatch}
 import com.nec.ve.VeColBatch
-import com.nec.ve.VeProcess.{OriginalCallingContext, calls}
-import com.nec.ve.VeRDD.{RichKeyedRDDL, RichRDD}
+import com.nec.ve.VeProcess.OriginalCallingContext
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.arrow.memory.BufferAllocator
 import org.apache.spark.TaskContext
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.expressions.Attribute
-import org.apache.spark.sql.execution.metric.SQLMetrics
 import org.apache.spark.sql.execution.{SparkPlan, UnaryExecNode}
 import org.apache.spark.sql.util.ArrowUtilsExposed
-
-import scala.concurrent.duration.NANOSECONDS
 
 object SparkToVectorEnginePlan {
 //  val ConvertColumnarToColumnar = false
@@ -43,7 +39,6 @@ case class SparkToVectorEnginePlan(childPlan: SparkPlan)
 
   override def executeVeColumnar(): RDD[VeColBatch] = {
     require(!child.isInstanceOf[SupportsVeColBatch], "Child should not be a VE plan")
-    import OriginalCallingContext.Automatic._
 
     // Instead of creating a new config we are reusing columnBatchSize. In the future if we do
     // combine with some of the Arrow conversion tools we will need to unify some of the configs.
@@ -58,8 +53,8 @@ case class SparkToVectorEnginePlan(childPlan: SparkPlan)
             implicit val allocator: BufferAllocator = ArrowUtilsExposed.rootAllocator
               .newChildAllocator(s"Writer for partial collector (ColBatch-->Arrow)", 0, Long.MaxValue)
             TaskContext.get().addTaskCompletionListener[Unit](_ => allocator.close())
-            import OriginalCallingContext.Automatic._
             import ImplicitMetrics._
+            import OriginalCallingContext.Automatic._
             if (ConvertColumnarToColumnar) {
               collectBatchMetrics(OUTPUT, ColumnarBatchToVeColBatch.toVeColBatchesViaCols(
                 columnarBatches = collectBatchMetrics(INPUT, columnarBatches),
